extract dispatcher wildcard tests
